Implement browse-by-mappings component in onion and standardize the suggestion, browse, and mappings-list components to use the same service. Fix bug where micromodule tag wasn't being highlighted correctly in various places. Fix bug where onion had two scroll bars and the context switcher would get pushed under the navbar. Closes #73. Closes #42. Closes #63.

+++ src/app/onion/learning-object-builder/suggestion/mappings-list.component.ts
@@ -1,5 +1,6 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter, SimpleChanges, OnChanges } from '@angular/core';
 import { SuggestionService } from './services/suggestion.service';
+import { OutcomeSuggestion } from '@cyber4all/clark-entity';
 
 @Component({
     selector: 'onion-mappings-list',
@@ -10,23 +11,18 @@ import { SuggestionService } from './services/suggestion.service';
 export class MappingsListComponent implements OnInit {
   // TODO: Add ability to deselect a mapped outcome
 
-  @Output('delete') delete: EventEmitter<string> = new EventEmitter<string>();
-  mappings;
-  showMappings = false;
+  @Input('mappings') mappings: Array<OutcomeSuggestion> = [];
 
   constructor(public service: SuggestionService) { }
 
   ngOnInit() {
     const m = this.service.mappings;
     this.mappings = m.currentValue;
-    this.showMappings = this.mappings.length >= 1;
-    m.observable.subscribe(data => {
-      this.mappings = data;
-      this.showMappings = this.mappings.length >= 1;
-    });
   }
 
   deleteMapping(m) {
-    this.service.removeMapping(m);
+    if (m && this.mappings.filter(x => x.id === m.id).length) {
+      this.service.removeMapping(m);
+    }
   }
 }
